# Frontend: Comprehensive code quality and accessibility improvements