--- a/frontend/src/components/common/ErrorBoundary.tsx
+++ b/frontend/src/components/common/ErrorBoundary.tsx
@@ -1,0 +1,114 @@
+import { Component, ErrorInfo, ReactNode } from 'react';
+import { AlertTriangle, RefreshCw } from 'lucide-react';
+
+interface ErrorBoundaryProps {
+  children: ReactNode;
+  fallback?: ReactNode;
+}
+
+interface ErrorBoundaryState {
+  hasError: boolean;
+  error: Error | null;
+  errorInfo: ErrorInfo | null;
+}
+
+export class ErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
+  constructor(props: ErrorBoundaryProps) {
+    super(props);
+    this.state = {
+      hasError: false,
+      error: null,
+      errorInfo: null,
+    };
+  }
+
+  static getDerivedStateFromError(error: Error): Partial<ErrorBoundaryState> {
+    return { hasError: true, error };
+  }
+
+  componentDidCatch(error: Error, errorInfo: ErrorInfo): void {
+    console.error('Error Boundary caught an error:', error, errorInfo);
+    this.setState({
+      error,
+      errorInfo,
+    });
+  }
+
+  handleReset = (): void => {
+    this.setState({
+      hasError: false,
+      error: null,
+      errorInfo: null,
+    });
+  };
+
+  render(): ReactNode {
+    if (this.state.hasError) {
+      if (this.props.fallback) {
+        return this.props.fallback;
+      }
+
+      return (
+        <div className="min-h-screen bg-gray-50 flex items-center justify-center p-4">
+          <div className="max-w-2xl w-full bg-white rounded-lg shadow-lg p-8">
+            <div className="flex items-center gap-4 mb-6">
+              <div className="bg-red-100 p-3 rounded-full">
+                <AlertTriangle className="text-red-600" size={32} />
+              </div>
+              <div>
+                <h1 className="text-2xl font-bold text-gray-900">
+                  Oops! Something went wrong
+                </h1>
+                <p className="text-gray-600 mt-1">
+                  We encountered an unexpected error. Please try refreshing the page.
+                </p>
+              </div>
+            </div>
+
+            {this.state.error && (
+              <div className="mb-6">
+                <div className="bg-red-50 border border-red-200 rounded-lg p-4">
+                  <h3 className="font-semibold text-red-900 mb-2">Error Details:</h3>
+                  <p className="text-sm text-red-800 font-mono break-all">
+                    {this.state.error.toString()}
+                  </p>
+                </div>
+
+                {import.meta.env.DEV && this.state.errorInfo && (
+                  <details className="mt-4">
+                    <summary className="cursor-pointer text-sm font-medium text-gray-700 hover:text-gray-900">
+                      Stack Trace (Development Only)
+                    </summary>
+                    <pre className="mt-2 text-xs bg-gray-50 p-4 rounded overflow-auto max-h-64 border">
+                      {this.state.errorInfo.componentStack}
+                    </pre>
+                  </details>
+                )}
+              </div>
+            )}
+
+            <div className="flex gap-4">
+              <button
+                onClick={this.handleReset}
+                className="flex items-center gap-2 px-6 py-3 bg-primary-600 text-white rounded-lg hover:bg-primary-700 transition-colors"
+                aria-label="Try again"
+              >
+                <RefreshCw size={20} />
+                Try Again
+              </button>
+              <button
+                onClick={() => window.location.href = '/'}
+                className="px-6 py-3 bg-gray-200 text-gray-800 rounded-lg hover:bg-gray-300 transition-colors"
+                aria-label="Go to homepage"
+              >
+                Go to Homepage
+              </button>
+            </div>
+          </div>
+        </div>
+      );
+    }
+
+    return this.props.children;
+  }
+}