… frontend/src/components/common/Header.tsx
… frontend/src/components/common/LoadingSpinner.tsx
… frontend/src/components/common/Sidebar.tsx
… frontend/src/index.css
… frontend/src/main.tsx